Bug Fixed.

+++ WindowsImpl/D3DRenderTarget.h
@@ -14,7 +14,7 @@ namespace Snowing::PlatformImpls::WindowsImpl::D3D
 
 		D3DTexture2D init(D3DTexture2D&& tex, const Handler& device);
 	public:
-		D3DRenderTarget(Math::Vec2<int> size);
+		D3DRenderTarget(Math::Vec2<size_t> size);
 
 		D3DRenderTarget(D3DTexture2D&& tex,const Handler& device);
 		D3DRenderTarget(D3DRenderTarget&&) = default;
